+++ project-model/src/main/java/com/walmartlabs/concord/project/model/Trigger.java
@@ -72,7 +72,7 @@ public class Trigger implements Serializable {
         if (cfg == null) {
             return false;
         }
-        Boolean v = (Boolean) cfg.get(Constants.Request.USE_INITIATOR);
+        Boolean v = (Boolean) cfg.get(Constants.Trigger.USE_INITIATOR);
         if (v == null) {
             return false;
         }